Delete register page
