+++ aural-front/app/components/loginHeader.tsx
@@ -2,8 +2,16 @@
 import React from 'react';
 import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
 import { router } from 'expo-router';
+import * as Linking from 'expo-linking';
+
 
 const LoginHeader = () => {
+  const handleRedirect = () => {
+      Linking.openURL('https://www.spotify.com/signup').catch(() => {
+        console.error('Failed to open Spotify signup.');
+      });
+    };
+
   const handleRegisterNavigation = () => {
     router.push("../registrationPage");
   };
@@ -12,8 +20,9 @@ const LoginHeader = () => {
     <View style={styles.container}>
       <Text style={styles.title}>Welcome</Text>
       <TouchableOpacity onPress={handleRegisterNavigation}>
-        <Text style={styles.registerText}>
-          Don't have an account? Register
+
+        <Text style={styles.redirectText} onPress={handleRedirect}>
+          Don't have a Spotify account? Register on Spotify
         </Text>
       </TouchableOpacity>
     </View>
@@ -37,5 +46,10 @@ const styles = StyleSheet.create({
     color: '#1DB954',
     textDecorationLine: 'underline',
     marginTop: 10,
+  },
+  redirectText: {
+    fontSize: 16,
+    color: '#1DB954',
+    textDecorationLine: 'underline',
   },
 });
